requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
PyPDF2
python-docx
python-multipart
python-dotenv
requests
cachetools
pymongo[srv]
certifi